requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sqlparse==0.5.3
urllib3==2.3.0
